ch8/pyproject.toml
[project]
name = "rag-with-haystack-ch8"
version = "0.1"
description = "Building pipelines with Haystack"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["llm", "rag", "nlp", "pipelines"]
authors = [
  {name = "Laura Funderburk"}
]

classifiers = [
  "Development Status :: 1 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "hayhooks>=1.5.0",
  "haystack-ai>=2.19.0",
  "ipykernel>=7.0.1",
  "ipywidgets>=8.1.8",
  "langgraph>=1.0.2",
  "nbformat>=5.10.4",
  "nltk>=3.9.1",
  "openpyxl>=3.1.5",
  "pandas>=2.3.3",
  "pymupdf>=1.26.5",
  "pypdf>=6.1.3",
  "pyspellchecker>=0.8.3",
  "pytest>=8.4.2",
  "python-docx>=1.2.0",
  "python-dotenv>=1.1.1",
  "python-pptx>=1.0.2",
  "rapidfuzz>=3.14.1",
  "requests>=2.32.0",
  "sentence-transformers>=5.0.0",
  "trafilatura>=2.0.0",
  "transformers[sentencepiece,torch]>=4.57.1",
]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"



[tool.pytest.ini_options]
addopts = "-v"
testpaths = [
  "pytests"
]

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff.lint]
select = [
    "A",
    "B",
    "D",
    "DTZ",
    "E",
    "EM",
    "F",
    "I",
    "PL",
    "W",
]
ignore = [
    "D105",
    "E731",
    "PLR",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"examples/*.py" = ["D"]
"test_*.py" = ["D"]
